requirements.txt
gradio
torch
torchvision
transformers
accelerate
bitsandbytes
pillow
sentencepiece
protobuf